+++ src/firecrest/status/health_check/health_checker_cluster.py
@@ -37,6 +37,7 @@ class ClusterHealthChecker:
                 settings.auth.authentication.public_certs,
                 username_claim=settings.auth.authentication.username_claim,
                 jwk_algorithm=settings.auth.authentication.jwk_algorithm,
+                audience=settings.auth.authentication.audience,
             )
         else:
             self.token_decoder = token_decoder